Reorganize `chalk_context::program_clauses`
